adding auth to all the requests

+++ src/main/java/com/example/journalApp/service/JournalEntryService.java
@@ -27,7 +27,7 @@ public class JournalEntryService {
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved =journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         } catch (Exception e) {
             System.out.println(e);
             throw new RuntimeException("something gone wrong",e);
@@ -43,12 +43,21 @@ public class JournalEntryService {
     public Optional<JournalEntry> findById(ObjectId id){
         return journalEntryRepository.findById(id);
     }
-    public void deleteById (ObjectId id, String userName){
+    @Transactional
+    public boolean deleteById (ObjectId id, String userName){
+        boolean removed = false;
+        try{
         User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
-
+        removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+        if(removed) {
+            userService.saveUser(user);
+            journalEntryRepository.deleteById(id);
+        }
+        } catch(Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleting the entry", e);
+        }
+        return removed;
     }
 
 }
